fix TeamCreateOutputPage props
